refactor(PostExecuteRequest): refactored to only one struct, added documentation
